# Backend/docs/openapi-final.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SIGGAB API",
    "version": "2.1.0",
    "description": "Sistema Gestor de Ganado Bovino (SIGGAB) - API REST.\n\nTodos los endpoints (excepto login y refresh) requieren Bearer token.\nRespuesta estandar: { success, data, message, errors }.",
    "contact": {
      "name": "SIGGAB Dev Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Desarrollo local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido en POST /auth/login"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "nullable": true
          },
          "message": {
            "type": "string",
            "example": "Operacion exitosa"
          },
          "errors": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "success",
          "data",
          "message",
          "errors"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/eventos-reproductivos": {
      "get": {
        "tags": [
          "ProductivoReproductivo"
        ],
        "summary": "Listar eventos reproductivos",
        "parameters": [
          {
            "in": "query",
            "name": "idAnimal",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "idLote",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "tipo",
            "schema": {
              "type": "string",
              "enum": [
                "CELO",
                "MONTA",
                "PRENEZ",
                "PARTO",
                "ABORTO"
              ]
            }
          },
          {
            "in": "query",
            "name": "estado",
            "schema": {
              "type": "string",
              "enum": [
                "PENDIENTE",
                "APROBADO",
                "RECHAZADO"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Eventos obtenidos"
          }
        }
      },
      "post": {
        "tags": [
          "ProductivoReproductivo"
        ],
        "summary": "Registrar evento reproductivo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "idAnimal",
                  "idLote",
                  "tipoEvento",
                  "fechaEvento"
                ],
                "properties": {
                  "idAnimal": {
                    "type": "integer"
                  },
                  "idLote": {
                    "type": "integer"
                  },
                  "tipoEvento": {
                    "type": "string",
                    "enum": [
                      "CELO",
                      "MONTA",
                      "PRENEZ",
                      "PARTO",
                      "ABORTO"
                    ]
                  },
                  "fechaEvento": {
                    "type": "string",
                    "format": "date"
                  },
                  "observaciones": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Evento creado"
          },
          "400": {
            "description": "Error de validacion"
          }
        }
      }
    },
    "/eventos-reproductivos/{id}": {
      "get": {
        "tags": [
          "ProductivoReproductivo"
        ],
        "summary": "Obtener evento reproductivo por ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento encontrado"
          },
          "404": {
            "description": "Evento no encontrado"
          }
        }
      },
      "patch": {
        "tags": [
          "ProductivoReproductivo"
        ],
        "summary": "Modificar evento reproductivo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tipoEvento": {
                    "type": "string",
                    "enum": [
                      "CELO",
                      "MONTA",
                      "PRENEZ",
                      "PARTO",
                      "ABORTO"
                    ]
                  },
                  "fechaEvento": {
                    "type": "string",
                    "format": "date"
                  },
                  "observaciones": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evento actualizado"
          },
          "400": {
            "description": "Error de validacion"
          },
          "404": {
            "description": "Evento no encontrado"
          }
        }
      }
    },
    "/eventos-reproductivos/{id}/validar": {
      "patch": {
        "tags": [
          "ProductivoReproductivo"
        ],
        "summary": "Validar evento reproductivo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "estadoValidacion"
                ],
                "properties": {
                  "estadoValidacion": {
                    "type": "string",
                    "enum": [
                      "APROBADO",
                      "RECHAZADO"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evento validado"
          },
          "400": {
            "description": "Error de validacion"
          },
          "404": {
            "description": "Evento no encontrado"
          }
        }
      }
    },
...
    "/reportes/productivo": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "summary": "Generar reporte productivo",
        "parameters": [
          {
            "in": "query",
            "name": "fechaInicio",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "fechaFin",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "idAnimal",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "idLote",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "formato",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv",
                "pdf"
              ],
              "default": "json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reporte generado (JSON, CSV o PDF)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              },
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}